Move IntlProvider file
